server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.6.0
sqlalchemy>=2.0.0
alibabacloud_docmind_api20220711>=1.3.0
alibabacloud_tea_openapi>=0.3.8
alibabacloud_tea_util>=0.3.11
alibabacloud_credentials>=0.3.2
PyYAML>=6.0

